refactor(app_user): move TicketEventMeta to logic/ — fix my_tickets→ticket cross-feature import (#2145)
TicketEventMeta is a pure data class with no feature dependencies.
Moves it from features/ticket/ui/model/ to logic/ (neutral layer),
eliminating:
- my_tickets/ui/my_tickets_page.dart: my_tickets → ticket
- home/logic/home_coordinator.dart: home → ticket (unlisted, fixed opportunistically)

All 5 callers updated to import from package:app_user/src/logic/ticket_event_meta.dart.
Also fixes pre-existing import ordering issues in app_routes.dart.

Constraint: no_cross_feature_imports lint rule prohibits features importing across feature boundaries
Confidence: high
Scope-risk: narrow
