+++ apps/app_user/lib/src/logic/ticket_event_meta.dart
